v.1

+++ internal/rest/deposits/entity.go
@@ -5,7 +5,7 @@ import "time"
 type Deposit struct {
 	ID        string    `gorm:"column:id;PRIMARY_KEY;NOT NULL"`
 	UserID    string    `gorm:"column:user_id;NOT NULL"`
-	Amount    float64   `gorm:"column:amount;NOT NULL"`
+	Amount    float32   `gorm:"column:amount;NOT NULL"`
 	Status    string    `gorm:"column:status;NOT NULL"`
 	CreatedAt time.Time `gorm:"column:created_at;NOT NULL"`
 	UpdatedAt time.Time `gorm:"column:updated_at;NOT NULL"`
